Test/Face_Tracking/requirements.txt

PySide6>=6.6
opencv-python>=4.10
mediapipe>=0.10
requests>=2.32
pyserial>=3.5
numpy>=1.26